feat: refactoring code and structure & 2 new components added
