+++ src/components/Coin/CoinLabel.jsx
@@ -1,0 +1,10 @@
+import React from "react";
+import PropTypes from 'prop-types';
+
+const CoinElement = ({ text }) => <p className="coin-label">{text}</p>
+
+CoinElement.propTypes = {
+  text: PropTypes.string
+}
+
+export default CoinElement;
